requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
streamlit>=1.40.0
crewai>=0.80.0
newsapi-python>=0.2.7
yfinance>=0.2.40
python-dotenv>=1.0.0
pydantic>=2.0.0
